rag_lambda_deployment_package/requirements.txt
langchain==0.1.20 # Use a specific version for stability
langchain-pinecone==0.1.0
langchain-community==0.0.38
langchain-aws==0.1.5
pinecone-client==3.2.2
boto3==1.34.108